Fixed room nav!
Ok, room nav now works.  Still needs a bit of debugging, but the new model has been implemented, which is basically that in Room's __init__, it makes a dictionary which holds the name of the door (back, a, and b), and for the values each one has the coords of the door, and an instance of Room().  Then, when the player enters that room, the next two rooms are generated.

Before, the rooms were being generated a bit too late I think, so some of the data couldn't be passed around.  Much easier to check if the player is standing on a door this way, and if they're standing on a door and type door, now room = that room.  Player position, merch data, everything is preserved.

yaaaay.