cleanup + load function
Whole buncha cleanup.

finally added in a load function that loads a presaved character file that I had to spend a fair amount f time tailoring.  I wonder if I could just write the character object as a binary file or something and read it back to load it, that would make it much easier.

regardless, it works, though I have a bit more testing to do, and I have one small fix- it's reading a character file in the main directory, not the saves.

After that I have to implement the save function, which should be pretty straight forward because I already have the code to write the toons __dict__ properly formated to a file that it may have to make.