Implementing save function
Just working out the last few bugs before the save and load functions work properly together.  Next will then be to test if it works when the file is updated: re, buy a skill, train a skill, buy some more inv items, save it, and see if it loads correctly, and then it's done!