save / load works
Now saving and loading works properly, even if you collect or sell more items.

The only problem is now with the skills.  The toon's skills are being stored as a class item in a dictionary.  But that doesn't transfer through the save as the save is all strings.  So I guess I just have to set the string skills to skill object instances after the character file is loaded.