save load works
Ok, should've saved this for one commit, note to sellf.

Save and load now works totally fine with skills.  You can train a skill, save, load it again and the current skill will be displayed still.

now, the skill system needs to be improved so that when a skill is finished training, it gets added to the toon.skills list, which is the attribute for trained, useable skills.  This will have to include a reset of the current_skill attribute to 'nope'.  Current setup should be able to handle that.