Skills fixed
Aquiring and training skills now works properly.

You buy a skillbook, which is a new class, from the merchant.  When you read the skillbook, you're given the option to train it.  You can only train one skill at a time, and if you overwrite a current skill that hasn't finished, you'll have to buy the book again.  I think that's ok for now.  When you train it, a check is performed at every run of the main while loop to see if it's finished, and if it has, it displays a message and resets current_skill to 'nope', adds the skill to the toon.skills list which will be the list that contains useable skills.  Probably have to be a dictionary of Skill items, the class of which will have to be defined.

Making the skills useable will require another push.